tests/categories//labels: move category and label factories to test folder, so they can be imported in other projects
