+++ tests/categories/conftest.py
@@ -1,6 +1,6 @@
 from pytest_factoryboy import register
 
-from tests.categories import factories as category_factories
+from adhocracy4.test.factories import categories as category_factories
 
 register(category_factories.CategoryFactory)
 register(category_factories.CategoryAliasFactory)
