+++ tests/labels/conftest.py
@@ -1,7 +1,7 @@
 from pytest_factoryboy import register
 
+from adhocracy4.test.factories import labels as label_factories
 from tests.apps.ideas import factories as idea_factories
-from tests.labels import factories as label_factories
 
 register(idea_factories.IdeaFactory)
 register(label_factories.LabelFactory)
